requirements.txt
discord.py==2.0.1
python-dotenv~=0.20.0
minio~=7.1.12